refactor(network): move to internal/network package
