+++ config.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"sync"
 
+	"github.com/jetkvm/kvm/internal/network"
 	"github.com/jetkvm/kvm/internal/usbgadget"
 )
 
@@ -73,27 +74,28 @@ func (m *KeyboardMacro) Validate() error {
 }
 
 type Config struct {
-	CloudURL             string             `json:"cloud_url"`
-	CloudAppURL          string             `json:"cloud_app_url"`
-	CloudToken           string             `json:"cloud_token"`
-	GoogleIdentity       string             `json:"google_identity"`
-	JigglerEnabled       bool               `json:"jiggler_enabled"`
-	AutoUpdateEnabled    bool               `json:"auto_update_enabled"`
-	IncludePreRelease    bool               `json:"include_pre_release"`
-	HashedPassword       string             `json:"hashed_password"`
-	LocalAuthToken       string             `json:"local_auth_token"`
-	LocalAuthMode        string             `json:"localAuthMode"` //TODO: fix it with migration
-	WakeOnLanDevices     []WakeOnLanDevice  `json:"wake_on_lan_devices"`
-	KeyboardMacros       []KeyboardMacro    `json:"keyboard_macros"`
-	EdidString           string             `json:"hdmi_edid_string"`
-	ActiveExtension      string             `json:"active_extension"`
-	DisplayMaxBrightness int                `json:"display_max_brightness"`
-	DisplayDimAfterSec   int                `json:"display_dim_after_sec"`
-	DisplayOffAfterSec   int                `json:"display_off_after_sec"`
-	TLSMode              string             `json:"tls_mode"` // options: "self-signed", "user-defined", ""
-	UsbConfig            *usbgadget.Config  `json:"usb_config"`
-	UsbDevices           *usbgadget.Devices `json:"usb_devices"`
-	DefaultLogLevel      string             `json:"default_log_level"`
+	CloudURL             string                 `json:"cloud_url"`
+	CloudAppURL          string                 `json:"cloud_app_url"`
+	CloudToken           string                 `json:"cloud_token"`
+	GoogleIdentity       string                 `json:"google_identity"`
+	JigglerEnabled       bool                   `json:"jiggler_enabled"`
+	AutoUpdateEnabled    bool                   `json:"auto_update_enabled"`
+	IncludePreRelease    bool                   `json:"include_pre_release"`
+	HashedPassword       string                 `json:"hashed_password"`
+	LocalAuthToken       string                 `json:"local_auth_token"`
+	LocalAuthMode        string                 `json:"localAuthMode"` //TODO: fix it with migration
+	WakeOnLanDevices     []WakeOnLanDevice      `json:"wake_on_lan_devices"`
+	KeyboardMacros       []KeyboardMacro        `json:"keyboard_macros"`
+	EdidString           string                 `json:"hdmi_edid_string"`
+	ActiveExtension      string                 `json:"active_extension"`
+	DisplayMaxBrightness int                    `json:"display_max_brightness"`
+	DisplayDimAfterSec   int                    `json:"display_dim_after_sec"`
+	DisplayOffAfterSec   int                    `json:"display_off_after_sec"`
+	TLSMode              string                 `json:"tls_mode"` // options: "self-signed", "user-defined", ""
+	UsbConfig            *usbgadget.Config      `json:"usb_config"`
+	UsbDevices           *usbgadget.Devices     `json:"usb_devices"`
+	NetworkConfig        *network.NetworkConfig `json:"network_config"`
+	DefaultLogLevel      string                 `json:"default_log_level"`
 }
 
 const configPath = "/userdata/kvm_config.json"
@@ -162,6 +164,10 @@ func LoadConfig() {
 
 	if loadedConfig.UsbDevices == nil {
 		loadedConfig.UsbDevices = defaultConfig.UsbDevices
+	}
+
+	if loadedConfig.NetworkConfig == nil {
+		loadedConfig.NetworkConfig = defaultConfig.NetworkConfig
 	}
 
 	config = &loadedConfig
