+++ jsonrpc.go
@@ -960,9 +960,10 @@ var rpcHandlers = map[string]RPCHandler{
 	"getDeviceID":            {Func: rpcGetDeviceID},
 	"deregisterDevice":       {Func: rpcDeregisterDevice},
 	"getCloudState":          {Func: rpcGetCloudState},
-	"getNetworkState":        {Func: rpcGetNetworkState},
-	"getNetworkSettings":     {Func: rpcGetNetworkSettings},
-	"renewDHCPLease":         {Func: rpcRenewDHCPLease},
+	"getNetworkState":        {Func: networkState.RpcGetNetworkState},
+	"getNetworkSettings":     {Func: networkState.RpcGetNetworkSettings},
+	"setNetworkSettings":     {Func: networkState.RpcSetNetworkSettings, Params: []string{"settings"}},
+	"renewDHCPLease":         {Func: networkState.RpcRenewDHCPLease},
 	"keyboardReport":         {Func: rpcKeyboardReport, Params: []string{"modifier", "keys"}},
 	"absMouseReport":         {Func: rpcAbsMouseReport, Params: []string{"x", "y", "buttons"}},
 	"relMouseReport":         {Func: rpcRelMouseReport, Params: []string{"dx", "dy", "buttons"}},
